refactor(config): remove the inert cache.disk_io_profile setting (#227)
The setting never affected I/O concurrency. Three separate places computed a
profile-based limit and nothing consumed any of them: the executor pool took a
flat DEFAULT_DISK_IO_CAPACITY, ResourcePoolConfig::with_disk_io_profile() had
no production callers, and the profile-sized StorageConcurrencyLimiter that
ServiceBuilder built sat behind #[allow(dead_code)]. Pool sizing now derives
from the executor's own capacities, which is the real constraint.

Removes the config key (with a DEPRECATED_KEYS entry so existing configs are
cleaned on upgrade), the setup wizard prompt that asked users to choose a
value that did nothing, the recommendation helper, both profile-based sizing
APIs and their constants, and ServiceBuilder's dead limiter field.

Also removes four StorageConcurrencyLimiter constructors with no references
anywhere, including tests: with_disk_io_defaults, with_cpu_defaults,
with_cpu_oversubscribe, for_disk_io_profile. Of seven constructors, one is
used in production.

Storage *detection* survives — it feeds the hardware report in
`xearthlayer setup`. With DiskIoProfile gone, config/storage.rs held nothing
but hardware detection, so it moves to system/storage_detect.rs returning
StorageType directly and the DiskIoProfile/StorageType pair collapses to one
type.

The live FUSE limiter at ortho_union_fs.rs:928, which gates every passthrough
read, is unchanged: it was never profile-based.

Refs #227
