+++ xearthlayer/src/config/mod.rs
@@ -37,7 +37,6 @@ mod keys;
 mod parser;
 mod settings;
 mod size;
-mod storage;
 mod texture;
 mod upgrade;
 mod writer;
@@ -100,10 +99,6 @@ pub use file::{
 };
 pub use keys::{ConfigKey, ConfigKeyError, SENSITIVE_VALUE_MASK};
 pub use size::{format_size, parse_size, Size, SizeParseError, GB, KB, MB};
-pub use storage::{
-    DiskIoProfile, DEFAULT_CPU_FALLBACK, HDD_IO_CEILING, HDD_IO_SCALING_FACTOR, NVME_IO_CEILING,
-    NVME_IO_SCALING_FACTOR, SSD_IO_CEILING, SSD_IO_SCALING_FACTOR,
-};
 pub use texture::TextureConfig;
 pub use upgrade::{
     analyze_config, upgrade_config, ConfigUpgradeAnalysis, UpgradeResult, DEPRECATED_KEYS,
