+++ xearthlayer/src/config/parser.rs
@@ -91,15 +91,6 @@ pub(super) fn parse_ini(ini: &Ini) -> Result<ConfigFile, ConfigFileError> {
                         reason: "expected a number between 0.0 and 1.0".to_string(),
                     })?;
         }
-        if let Some(v) = section.get("disk_io_profile") {
-            config.cache.disk_io_profile =
-                v.parse().map_err(|_| ConfigFileError::InvalidValue {
-                    section: "cache".to_string(),
-                    key: "disk_io_profile".to_string(),
-                    value: v.to_string(),
-                    reason: "must be one of: auto, hdd, ssd, nvme".to_string(),
-                })?;
-        }
     }
 
     // [texture] section
