+++ xearthlayer/src/executor/mod.rs
@@ -156,8 +156,8 @@ pub use handle::{JobHandle, JobStatus, Signal};
 pub use resource_pool::{
     ResourcePermit, ResourcePool, ResourcePoolConfig, ResourcePools, ResourceType,
     DEFAULT_CPU_CAPACITY_MULTIPLIER, DEFAULT_DISK_IO_CAPACITY, DEFAULT_MAX_PREFETCH_FRACTION,
-    DEFAULT_NETWORK_CAPACITY_MULTIPLIER, DISK_IO_CAPACITY_HDD, DISK_IO_CAPACITY_NVME,
-    DISK_IO_CAPACITY_SSD, MAX_NETWORK_CAPACITY, MIN_CPU_CAPACITY_ADDITION, MIN_NETWORK_CAPACITY,
+    DEFAULT_NETWORK_CAPACITY_MULTIPLIER, MAX_NETWORK_CAPACITY, MIN_CPU_CAPACITY_ADDITION,
+    MIN_NETWORK_CAPACITY,
 };
 
 // Telemetry
